+++ src/Ext/Techno/Body.Update.cpp
@@ -381,15 +381,16 @@ void TechnoExt::ExtData::ApplySpawnLimitRange()
 	}
 }
 
-void TechnoExt::ExtData::UpdateTypeData(TechnoTypeClass* currentType)
+void TechnoExt::ExtData::UpdateTypeData(TechnoTypeClass* pCurrentType)
 {
 	auto const pThis = this->OwnerObject();
-	auto const oldType = this->TypeExtData;
+	auto const pOldTypeExt = this->TypeExtData;
+	auto const pOldType = this->TypeExtData->OwnerObject();
 
 	if (this->LaserTrails.size())
 		this->LaserTrails.clear();
 
-	this->TypeExtData = TechnoTypeExt::ExtMap.Find(currentType);
+	this->TypeExtData = TechnoTypeExt::ExtMap.Find(pCurrentType);
 
 	// Recreate Laser Trails
 	for (auto const& entry : this->TypeExtData->LaserTrailData)
@@ -410,20 +411,49 @@ void TechnoExt::ExtData::UpdateTypeData(TechnoTypeClass* currentType)
 		this->PassengerDeletionTimer.Stop();
 
 	// Remove from tracked AutoDeath objects if no longer has AutoDeath
-	if (oldType->AutoDeath_Behavior.isset() && !this->TypeExtData->AutoDeath_Behavior.isset())
+	if (pOldTypeExt->AutoDeath_Behavior.isset() && !this->TypeExtData->AutoDeath_Behavior.isset())
 	{
 		auto& vec = HouseExt::ExtMap.Find(pThis->Owner)->OwnedAutoDeathObjects;
 		vec.erase(std::remove(vec.begin(), vec.end(), this), vec.end());
 	}
 
-	auto const rtti = oldType->OwnerObject()->WhatAmI();
+	auto const rtti = pOldType->WhatAmI();
 
 	// Remove from limbo reloaders if no longer applicable
 	if (rtti != AbstractType::AircraftType && rtti != AbstractType::BuildingType
-		&& oldType->OwnerObject()->Ammo > 0 && oldType->ReloadInTransport && !this->TypeExtData->ReloadInTransport)
+		&& pOldType->Ammo > 0 && pOldTypeExt->ReloadInTransport && !this->TypeExtData->ReloadInTransport)
 	{
 		auto& vec = HouseExt::ExtMap.Find(pThis->Owner)->OwnedTransportReloaders;
 		vec.erase(std::remove(vec.begin(), vec.end(), this), vec.end());
+	}
+
+	// Update open topped state of potential passengers if transport's OpenTopped value changes.
+	bool toOpenTopped = pCurrentType->OpenTopped && !pOldType->OpenTopped;
+
+	if ((toOpenTopped || (!pCurrentType->OpenTopped && pOldType->OpenTopped)) && pThis->Passengers.NumPassengers > 0)
+	{
+		auto pPassenger = pThis->Passengers.FirstPassenger;
+
+		while (pPassenger)
+		{
+			if (toOpenTopped)
+			{
+				pThis->EnteredOpenTopped(pPassenger);
+			}
+			else
+			{
+				pThis->ExitedOpenTopped(pPassenger);
+
+				// Lose target & destination
+				pPassenger->Guard();
+
+				// OpenTopped adds passengers to logic layer when enabled. Under normal conditions this does not need to be removed since
+				// OpenTopped state does not change while passengers are still in transport but in case of type conversion that can happen.
+				MapClass::Logics.get().RemoveObject(pPassenger);
+			}
+
+			pPassenger = abstract_cast<FootClass*>(pPassenger->NextObject);
+		}
 	}
 }